+++ src/main/java/com/example/DDP/Movimientos/Repository/MovClientesRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MovClientesRepository extends CrudRepository<MovimientosCliente, Long> {
+public interface MovClientesRepository extends CrudRepository<MovimientosCliente, Integer> {
 
 }
